Using organization feature of latest keycloak version
Keycloak has a "new" feature called organization for multi-tenancy. I'm changing the project to use this feature and assign the customer entity to the organization that the authenticated user belongs. Due to these changes, I'm moving the common code used in the authentication flow to another lib called auth-lib that can be re-used and better cohesion.