Develop (#9)
Using the Organization Feature of the Latest Keycloak Version

Keycloak has introduced a "new" feature called Organization for multi-tenancy. I am updating the project to utilize this feature and assign the customer entity to the organization that the authenticated user belongs to. As part of these changes, I am moving the common code used in the authentication flow to a separate library called auth-lib, which can be reused for better cohesion.

Additionally, I am adding Hibernate's second-level cache and integrating Spring Cache using Redis.